WE-23 SocketEventProvider 생성
기존 Auth 에서 하던 작업을 SocketEventProvider 로 옮깁니다!
